Stockton (#1)
Update for the Stockton.  This release adds a settings button that allows the user to toggle between light and dark themes.  It also replaces static data in the Capability Assessment Page with dynamic data from the database and adds some reactivity to some of the controls.  The chip in the list of capabilities now brings up detail box.